chore: Use type-safe invalidateQueries
