+++ application/ui/src/query-client/query-client.ts
@@ -1,0 +1,69 @@
+// Copyright (C) 2025-2026 Intel Corporation
+// SPDX-License-Identifier: Apache-2.0
+
+import { toast } from '@geti/ui';
+import { MutationCache, QueryClient } from '@tanstack/react-query';
+
+import { paths } from '../api/openapi-spec';
+import { Meta, QueryKey } from './query-client.interface';
+
+declare module '@tanstack/react-query' {
+    interface Register {
+        mutationMeta: Meta;
+        queryMeta: Meta;
+    }
+}
+
+const TOAST_DURATION = 5000;
+
+const getErrorMessage = (error: unknown): string => {
+    if (error && typeof error === 'object') {
+        if ('detail' in error && typeof error.detail === 'string') {
+            return error.detail;
+        }
+
+        if ('message' in error && typeof error.message === 'string') {
+            if (error instanceof TypeError && error.message.includes('Failed to fetch')) {
+                return 'Network error. Please check your connection and try again.';
+            }
+
+            return error.message;
+        }
+    }
+
+    return 'An unexpected error occurred. Please try again.';
+};
+
+export const queryClient = new QueryClient({
+    defaultOptions: {
+        queries: {
+            retry: false,
+        },
+        mutations: {
+            retry: false,
+        },
+    },
+    mutationCache: new MutationCache({
+        onSuccess: (_data, _variables, _context, mutation) => {
+            const meta = mutation.meta;
+            const invalidateQueries = meta?.invalidateQueries;
+
+            if (invalidateQueries) {
+                invalidateQueries.forEach((queryKey) => {
+                    queryClient.invalidateQueries({ queryKey });
+                });
+            }
+        },
+        onError: (error) => {
+            toast({
+                type: 'error',
+                message: getErrorMessage(error),
+                duration: TOAST_DURATION,
+            });
+        },
+    }),
+});
+
+export const getQueryKey = (queryKey: QueryKey<paths>): QueryKey<paths> => {
+    return queryKey;
+};
